+++ src/main/java/com/example/newsfeed/member/controller/AuthenticationController.java
@@ -1,7 +1,7 @@
 package com.example.newsfeed.member.controller;
 
 import com.example.newsfeed.global.entity.SessionMemberDto;
-import com.example.newsfeed.member.dto.LoginRequestDto;
+import com.example.newsfeed.member.dto.request.LoginRequestDto;
 import com.example.newsfeed.member.entity.Member;
 import com.example.newsfeed.member.service.MemberService;
 import jakarta.servlet.http.HttpServletRequest;
