+++ src/main/java/com/example/newsfeed/member/dto/request/LoginRequestDto.java
@@ -1,10 +1,12 @@
-package com.example.newsfeed.member.dto;
+package com.example.newsfeed.member.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
+@AllArgsConstructor
 public class LoginRequestDto {
 
     @Pattern(regexp = "^[\\w!#$%&'*+/=?`{|}~^.-]+@[\\w.-]+\\.[a-zA-Z]{2,6}$",
