+++ src/main/java/com/example/newsfeed/member/dto/request/MemberDeleteRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.newsfeed.member.dto;
+package com.example.newsfeed.member.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
@@ -6,7 +6,7 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class deleteRequestDto {
+public class MemberDeleteRequestDto {
 
     @NotBlank(message = "비밀번호를 입력해주세요")
     private final String password;
