+++ src/main/java/com/example/newsfeed/member/service/MemberService.java
@@ -5,7 +5,7 @@ import com.example.newsfeed.global.entity.SessionMemberDto;
 import com.example.newsfeed.member.dto.request.MemberRequestDto;
 import com.example.newsfeed.member.dto.response.MemberResponseDto;
 import com.example.newsfeed.member.dto.response.MemberUpdateProfileResponseDto;
-import com.example.newsfeed.member.dto.updatePasswordRequestDto;
+import com.example.newsfeed.member.dto.request.MemberUpdatePasswordRequestDto;
 import com.example.newsfeed.member.dto.request.MemberUpdateProfileRequestDto;
 import com.example.newsfeed.member.entity.Member;
 import com.example.newsfeed.member.repository.MemberRepository;
@@ -22,6 +22,16 @@ import java.util.Optional;
 public class MemberService {
 
     private final MemberRepository memberRepository;
+
+    public Member loginMember(String email, String password) {
+
+        Member findMember = findActiveMemberByEmailOrElseThrow(email);
+
+        if (!PasswordEncoder.matches(password, findMember.getPassword())) {
+            throw new RuntimeException("비밀번호가 불일치");
+        }
+        return findMember;
+    }
 
     @Transactional
     public MemberResponseDto createMember(MemberRequestDto dto) {
@@ -72,9 +82,9 @@ public class MemberService {
     }
 
     @Transactional
-    public void updatePassword(SessionMemberDto sessionMemberDto, updatePasswordRequestDto dto) {
+    public void updatePassword(SessionMemberDto session, MemberUpdatePasswordRequestDto dto) {
 
-        Member findMember = findActiveMemberByEmailOrElseThrow(sessionMemberDto.getEmail());
+        Member findMember = findActiveMemberByEmailOrElseThrow(session.getEmail());
 
         /*본인 확인을 위해 입력한 현재 비밀번호가 일치하지 않은 경우*/
         if(!PasswordEncoder.matches(dto.getOldPassword(), findMember.getPassword())) {
@@ -93,27 +103,17 @@ public class MemberService {
 
         /*비밀번호 암호화 후 업데이트*/
         findMember.updatePassword(PasswordEncoder.encode(dto.getNewPassword()));
-
         log.info("비밀번호 수정 성공");
     }
 
     @Transactional
-    public void deleteMember(SessionMemberDto sessionMemberDto, String password) {
-        Member findMember = findActiveMemberByEmailOrElseThrow(sessionMemberDto.getEmail());
+    public void deleteMember(SessionMemberDto session, String password) {
+        Member findMember = findActiveMemberByEmailOrElseThrow(session.getEmail());
 
         if (!PasswordEncoder.matches(password, findMember.getPassword())) {
             throw new RuntimeException("입력받은 비밀번호와 유저의 비밀번호가 다름");
         }
-
-        // hard delete 대신 isDelete 를 true 하는 방식으로 soft delete
-        // 삭제된 유저 조회시 "삭제된 유저입니다" 라고 출력될 수 있도록 nickname 필드를 더티체킹
         findMember.updateIsDeleteTrueAndNickname();
-    }
-
-    private Member findMemberByEmailOrElseThrow(String email) {
-        return memberRepository.findMemberByEmail(email).orElseThrow(() ->
-                new RuntimeException("찾아지는 이메일 유저가 없음")
-        );
     }
 
     private Member findActiveMemberByEmailOrElseThrow(String email) {
@@ -121,25 +121,8 @@ public class MemberService {
                 new RuntimeException("탈퇴하지 않은 유저들 중에 찾아지는 이메일 유저가 없음"));
     }
 
-    public Member findMemberByIdOrElseThrow(Long memberId) {
-        return memberRepository.findMemberById(memberId).orElseThrow(() ->
-                new RuntimeException("찾아지는 아이디 유저가 없음")
-        );
-    }
-
     public Member findActiveMemberByIdOrElseThrow(Long id) {
         return memberRepository.findActiveMemberById(id).orElseThrow(() ->
                 new RuntimeException("탈퇴하지 않은 유저들 중에 찾아지는 id 유저가 없음"));
-    }
-
-    public Member loginMember(String email, String password) {
-
-        Member findMember = findActiveMemberByEmailOrElseThrow(email);
-
-        if (!PasswordEncoder.matches(password, findMember.getPassword())) {
-            throw new RuntimeException("비밀번호가 불일치");
-        }
-
-        return findMember;
     }
 }
